MS Windows CHOICE command, (yes/no)
Demonstration program showing to use the MS Windows DOS 'CHOICE' command. This can be used in a scenario where the user has to answer a question with 'yes' or 'no' before the program will continue. This is particularly useful if the batch script is about to do something potentially destructive or have a serious impact on something.