File History user guide
File History is a feature in Microsoft Windows operating system version 10 that automatically creates backups of files that are present in folders. This ensures that even if the original files are lost, damaged or deleted, you can still restore them with ease. Moreover, you can also access different versions of your files from a specific point in time as per your requirement. This guide was written with Microsoft Windows version 10 in mind.